Improved CSV export to better comply with the existing Arctic structure
